% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate function.R
\name{allocate}
\alias{allocate}
\title{Allocate Function}
\usage{
allocate(data, none, software)
}
\arguments{
\item{data}{conjoint cho file. Required}

\item{none}{Whether none concept is included or not in design. Requires a value of 0 for not included or 1 for included}

\item{software}{Specify if allocation should be for CBC/HB or ChoiceModelR. Requires a value of 0 for ChoiceModelR or 1 for CBC/HB}
}
\value{
allocated cho file
}
\description{
This function creates allocated cho file.
}
