% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatcho function.R
\name{createcho}
\alias{createcho}
\title{Create Cho Function}
\usage{
createcho(design, data, rating)
}
\arguments{
\item{design}{conjoint design file. Required}

\item{data}{conjoint data file. Required}

\item{rating}{likelihood ratings after exercise. Optional}
}
\value{
cho file
}
\description{
This function creates cho file from merging conjoint design with data. If ratings are available, they will be merged into the file as well.
}
